Move mock_aws_request_headers to localstack.utils.aws.headers
