+++ localstack/services/apigateway/integration.py
@@ -42,7 +42,6 @@ from localstack.services.apigateway.templates import (
 )
 from localstack.services.stepfunctions.stepfunctions_utils import await_sfn_execution_result
 from localstack.utils import common
-from localstack.utils.aws import aws_stack
 from localstack.utils.aws.arns import extract_region_from_arn
 from localstack.utils.aws.aws_responses import (
     LambdaResponse,
@@ -50,6 +49,7 @@ from localstack.utils.aws.aws_responses import (
     requests_response,
 )
 from localstack.utils.aws.client_types import ServicePrincipal
+from localstack.utils.aws.request_context import mock_aws_request_headers
 from localstack.utils.aws.templating import VtlTemplate
 from localstack.utils.collections import dict_multi_values, remove_attributes
 from localstack.utils.common import make_http_request, to_str
@@ -151,7 +151,7 @@ def get_internal_mocked_headers(
         )
     else:
         access_key_id = None
-    headers = aws_stack.mock_aws_request_headers(
+    headers = mock_aws_request_headers(
         service=service_name, aws_access_key_id=access_key_id, region_name=region_name
     )
 
@@ -589,7 +589,7 @@ class S3Integration(BackendIntegration):
             LOG.debug(msg)
             return make_error_response(msg, 404)
 
-        headers = aws_stack.mock_aws_request_headers(
+        headers = mock_aws_request_headers(
             service="s3",
             aws_access_key_id=invocation_context.account_id,
             region_name=invocation_context.region_name,
@@ -715,7 +715,7 @@ class SNSIntegration(BackendIntegration):
             LOG.warning("Failed to apply template for SNS integration", e)
             raise
         region_name = uri.split(":")[3]
-        headers = aws_stack.mock_aws_request_headers(
+        headers = mock_aws_request_headers(
             service="sns", aws_access_key_id=invocation_context.account_id, region_name=region_name
         )
         result = make_http_request(
@@ -780,7 +780,7 @@ class StepFunctionIntegration(BackendIntegration):
         result = json_safe(remove_attributes(result, ["ResponseMetadata"]))
         response = StepFunctionIntegration._create_response(
             HTTPStatus.OK.value,
-            aws_stack.mock_aws_request_headers(
+            mock_aws_request_headers(
                 "stepfunctions",
                 aws_access_key_id=invocation_context.account_id,
                 region_name=invocation_context.region_name,
