+++ tests/aws/services/kinesis/test_kinesis.py
@@ -11,7 +11,8 @@ from localstack import config, constants
 from localstack.constants import TEST_AWS_ACCESS_KEY_ID, TEST_AWS_REGION_NAME
 from localstack.services.kinesis import provider as kinesis_provider
 from localstack.testing.pytest import markers
-from localstack.utils.aws import aws_stack, resources
+from localstack.utils.aws import resources
+from localstack.utils.aws.request_context import mock_aws_request_headers
 from localstack.utils.common import poll_condition, retry, select_attributes, short_uid
 from localstack.utils.kinesis import kinesis_connector
 
@@ -250,7 +251,7 @@ class TestKinesis:
         # get records with CBOR encoding
         iterator = get_shard_iterator(stream_name, aws_client.kinesis)
         url = config.internal_service_url()
-        headers = aws_stack.mock_aws_request_headers(
+        headers = mock_aws_request_headers(
             "kinesis", aws_access_key_id=TEST_AWS_ACCESS_KEY_ID, region_name=TEST_AWS_REGION_NAME
         )
         headers["Content-Type"] = constants.APPLICATION_AMZ_CBOR_1_1
@@ -284,7 +285,7 @@ class TestKinesis:
 
         # empty get records with CBOR encoding
         url = config.internal_service_url()
-        headers = aws_stack.mock_aws_request_headers(
+        headers = mock_aws_request_headers(
             "kinesis", aws_access_key_id=TEST_AWS_ACCESS_KEY_ID, region_name=TEST_AWS_REGION_NAME
         )
         headers["Content-Type"] = constants.APPLICATION_AMZ_CBOR_1_1
