+++ tests/aws/services/sts/test_sts.py
@@ -8,7 +8,7 @@ from localstack import config
 from localstack.constants import APPLICATION_JSON, TEST_AWS_ACCESS_KEY_ID, TEST_AWS_REGION_NAME
 from localstack.testing.aws.util import create_client_with_keys
 from localstack.testing.pytest import markers
-from localstack.utils.aws import aws_stack
+from localstack.utils.aws.request_context import mock_aws_request_headers
 from localstack.utils.numbers import is_number
 from localstack.utils.strings import short_uid, to_str
 
@@ -177,7 +177,7 @@ class TestSTSIntegrations:
     def test_expiration_date_format(self):
         url = config.internal_service_url()
         data = {"Action": "GetSessionToken", "Version": "2011-06-15"}
-        headers = aws_stack.mock_aws_request_headers(
+        headers = mock_aws_request_headers(
             "sts", aws_access_key_id=TEST_AWS_ACCESS_KEY_ID, region_name=TEST_AWS_REGION_NAME
         )
         headers["Accept"] = APPLICATION_JSON
