+++ tests/unit/test_dynamodb.py
@@ -10,8 +10,8 @@ from localstack.services.dynamodb.utils import (
     SchemaExtractor,
     dynamize_value,
 )
-from localstack.utils.aws import aws_stack
 from localstack.utils.aws.arns import dynamodb_table_arn
+from localstack.utils.aws.request_context import mock_aws_request_headers
 
 
 def test_fix_region_in_headers():
@@ -19,7 +19,7 @@ def test_fix_region_in_headers():
     # TODO: this may need to be updated once we migrate DynamoDB to ASF
 
     for region_name in ["local", "localhost"]:
-        headers = aws_stack.mock_aws_request_headers(
+        headers = mock_aws_request_headers(
             "dynamodb", aws_access_key_id=TEST_AWS_ACCESS_KEY_ID, region_name=region_name
         )
         assert TEST_AWS_REGION_NAME not in headers.get("Authorization")
